Delete the ping timer when the bridge's connection is closed.
When a bridge is created, the ping loop is started right away. This
happens whether or not the bridge is ever actually used. In a perfect
world, I'd refactor this so that the ping loop is only started if the
bridge is ever used.

For now, this change adds a helper method to close the connection and
also delete the timer that runs the ping loop. After this, the client
should be garbage collected.