Add support for Buttons, fix client singleton promise, minor fixes
This change adds support for Button-related messages, including
Button-related classes. This is largely just classes in the message body
types that are then used by `homebridge-lutron-caseta-leap`. There are
also a few helper methods added to `SmartBridge` to get buttongroups and
buttons.

This commit also changes how `LeapClient` connections work, hopefully
fixing the dumb race condition. This doesn't change the contract, but
uses a promise as a connected indicator, and users should await that.

This also contains some lint fixes and prettier changes.