Return an ExceptionDetail from getButtonGroupsFromDevice if appropriate
This change permits the SmartBridge::getButtonGroupsFromDevice method to
return ExceptionDetail objects in addition to/instead of the ButtonGroup
objects, if the bridge returns a 404 upon requesting it/them. This can
happen when a Pico has been removed from the bridge. This should
prooooobably get solved in a more general way, since anything can get
removed and the code should support that.