Add support for bridge pairing
BREAKING CHANGE: This change alters how the BridgeFinder class works.
This library no longer concerns itself with how secrets are handled, nor
with constructing bridge objects. This is to enable the BridgeFinder to
be used for other purposes (like pre-pairing discovery) and to enable
other bridge types to be supported. Now users will get network info
about the discovered bridge, and they can construct a client (or not) as
they require.

This change also adds support for pairing of Caseta Smart Bridge 2
devices. A special-purpose, not-quite-LEAP client is added just for
pairing, along with required key material.