Refactor the ping loop; start the timer after writing
This change refactors the SmartBridge ping loop to make it simpler and
behave more like a keepalive. Now we don't particularly care if a ping
fails because future requests (either the next ping, or external calls)
will trigger the client to reconnect (and block until it does,
potentially timing *those* calls out). We will keep trying to reconnect
with every ping interval until told to stop.

This also tweaks how the LeapClient request method handles request
response timeouts. Previously, we would start the clock when we
submitted data to the socket, not when it was successfully written. This
means that slow writes (for whatever reason) would count against the
time that the bridge had to respond. Now, we don't start that timer
until we know the data's been written.

This change also makes it a lot more explicit that we delete in-flight
requests (including their timeout) if the provided tag exists.
Basically, if you re-use a tag, it clobbers the old request, and rejects
it.

Lastly, this commit makes pings less-frequent, happening every 5 minutes
intead of every minute.