Don't re-request an SRV record
Refactors the code so as not to repeat a SRV request for information
about the subject of the PTR mDNS broadcast query. This should help
eliminate avenues for confusion when we get a PTR for bridge 1, ask for
SRV, and bridge 2 gets back to us more quickly. Or problems that _would_
arise, were we not doing broken (unicast) requests that some network
hardware will not forward, causing user reports.

So anyway, this should help the problem of network hardware not
forwarding mDNS requests, leading to bridge discovery not working, and
also a thing that hadn't yet become a problem.

Know that tinkerhub-mdns' mDNS manager fetches PTR records by default
when making search broadcast queries. It'll also get SRV requests, but
since there's no actual interface to acquire the manager in order to
reach that SRV request response, we do some ugly things to get at it. I
dunno, man, it works on my box.