Support version checking, CommuniqueTypes, and SSL logging
This change adds explicit enums for CommuniqueTypes, instead of
otherwise-unconstrained strings.

This also adds support for the `clientsetting` URL, which enables client
users to check LEAP versions when communicating with a bridge.

Lastly, this enables the LeapClient to log SSL key data for debugging.
This is done by passing a file stream writer to the client.