Add a bunch of new message body types, retrive method, emit disconnected
This change adds a bunch of new body types (largely) in support of
virtualbuttons (read: scenes).

This also adds a `retrieve` method to the LeapClient. This is a generic
method that should have existed a loooooong time ago, but anyway now it
does.

Lastly, this change makes SmartBridge also emit disconnected event.